Server/server/requirements.txt
flask
sqlalchemy
psycopg2-binary
pynacl